requirements.txt
google-cloud-monitoring
google-cloud-monitoring-dashboards
google-cloud-monitoring-metrics-scopes